requirements.txt
Jinja2==2.7.1
MarkupSafe==0.18
PyDispatcher==2.0.3
Twisted==13.1.0
pymodbus==1.1.0
pyserial==2.6
python-dateutil==2.1
six==1.4.1
wsgiref==0.1.2
zope.interface==4.0.3
transitions
pymdbpc
